backend/requirements.txt
fastapi==0.110.0
uvicorn==0.30.3
SQLAlchemy==2.0.29
pydantic==1.10.14
python-dotenv==1.0.1
APScheduler==3.10.4
google-generativeai==0.1.0  # or the pinned version you installed
httpx==0.27.0


